Separate fake provider implementation.
